pyproject.toml
[tool.poetry]
name = "nen2660"
version = "0.9.0"
description = ""
authors = ["Redmer Kronemeijer <redmer.kronemeijer@crow.nl>"]

packages = [
    { include = "rdflibplugins", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
rdflib = "^6.0.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
